requirements.txt
Flask
tensorflow
numpy
pillow
opencv-python
